backend/python-agent/pyproject.toml
[tool.poetry]
name = "agent-service-py"
version = "0.1.0"
description = "AutoAgent Agent Session Service for executing tasks."
authors = ["AutoAgent Core Team <dev@autoagent.is>"]
license = "Proprietary"

packages = [{ include = "agent_service_py", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
# Note: autoagent-api will be installed from wheel in Docker build
# Core libraries
grpcio = "^1.60.0"
python-dotenv = "^1.0.0"
bufbuild-protovalidate-protocolbuffers-pyi = "31.1.0.1.20250612022732+297b8109523d"
bufbuild-protovalidate-protocolbuffers-python = "31.1.0.1.20250612022732+297b8109523d"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.60.0"
types-protobuf = "^6.30.0"
mypy = "^1.8.0"
pytest = "^8.0.0"

[tool.poetry.group.test.dependencies]
# Tests de base
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-xdist = "^3.5.0"    # Tests parallèles
pytest-cov = "^4.0.0"      # Couverture de code

# Tests gRPC (approche classique)
grpcio-testing = "^1.60.0" # Framework officiel Google
pytest-grpc = "^0.8.0"     # Plugin pytest pour gRPC

# Tests Docker (approche moderne)
pytest-docker = "^3.1.0"  # Tests Docker natifs
testcontainers = "^4.0.0" # Conteneurs de test isolés

# Tests Buf/Protobuf
protobuf = "^6.31.1" # Pour manipulation messages

[[tool.poetry.source]]
name = "buf"
url = "https://buf.build/gen/python"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
